Adding a version serialization type to ConvPackedParam
ghstack-source-id: 9576ad140435f39eacf5a63ccb58dda62fee9e36
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43086

[rfc] temporary custom pickle functions for maintaining BC on ConvPackedParams

Summary:

This is an another approach to changing ConvPackedParams structure
while maintaining BC.  We add the ability to add custom pickle functions
where `__set_state__` takes a `c10::IValue` and knows how to parse it,
and define the custom pickle methods with the parsing logic on
`ConvPackedParams`.

There are various TODOs before land left in the diff, ideally we
can align if this is the approach people are OK with and implement
the TODOs after consensus.

For additional context, https://github.com/pytorch/pytorch/pull/43087
contained an alternative version of this.

Test Plan:

For now, check that saving a model with
https://gist.github.com/vkuzo/f3616c5de1b3109cb2a1f504feed69be
on an old revision and loading it on this revision works.
Tests would be needed if this approach sounds good.

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 9576ad140435f39eacf5a63ccb58dda62fee9e36
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43230